correct name of method for fetching FQCN
the current method does not seem to exist, but this one will return the
FQCN of the metadata being wrapped, which seems to be what it intends to
do.

there are no tests for this code atm, and the ClassMetadata service has
properties accessed which do not exist on the interface which makes
mocking or adding a dummy problematic. static analysis would also have
trouble here I think.